+++ src/components/DownloadAndShareComponents/NoResourcesToShare/NoResourcesToShare.test.tsx
@@ -1,0 +1,11 @@
+import renderWithTheme from "../../../__tests__/__helpers__/renderWithTheme";
+
+import NoResourcesToShare from ".";
+
+describe("NoResourcesToDownload", () => {
+  it("renders NoResourcesToShare component", () => {
+    const { getByText } = renderWithTheme(<NoResourcesToShare />);
+
+    expect(getByText("No resources to share")).toBeInTheDocument();
+  });
+});