+++ src/components/DownloadAndShareComponents/downloadsAndShare.types.ts
@@ -3,6 +3,18 @@ import { z } from "zod";
 import { ExpandingContainerTitle } from "../ExpandingContainer/ExpandingContainer";
 import { LessonItemTitle } from "../LessonItemContainer/LessonItemContainer";
 
+import {
+  LessonDownloadsData,
+  LessonShareData,
+  LessonShareSchema,
+} from "@/node-lib/curriculum-api";
+
+export type Resources =
+  | LessonDownloadsData["downloads"]
+  | LessonShareData["shareableResources"];
+
+export type ResourceType = LessonShareSchema["type"] | DownloadResourceType;
+
 const DOWNLOAD_RESOURCE_TYPES = [
   "presentation",
   "intro-quiz-questions",
@@ -15,6 +27,16 @@ const DOWNLOAD_RESOURCE_TYPES = [
   "supplementary-docx",
 ] as const;
 
+export const preselectedResourceType = z.union([
+  z.literal("slide deck"),
+  z.literal("exit quiz"),
+  z.literal("starter quiz"),
+  z.literal("worksheet"),
+  z.literal("additional material"),
+  z.literal("video"),
+  z.literal("all"),
+]);
+
 export const preselectedDownloadType = z.union([
   z.literal("slide deck"),
   z.literal("exit quiz"),
@@ -24,7 +46,27 @@ export const preselectedDownloadType = z.union([
   z.literal("all"),
 ]);
 
-export type PreselectedDownloadType = z.infer<typeof preselectedDownloadType>;
+export const preselectedShareType = z.union([
+  z.literal("exit quiz"),
+  z.literal("starter quiz"),
+  z.literal("worksheet"),
+  z.literal("video"),
+  z.literal("all"),
+]);
+
+export type PreselectedResourceType = z.infer<typeof preselectedResourceType>;
+
+export type PreselectedDownloadType = Exclude<
+  z.infer<typeof preselectedResourceType>,
+  "video"
+>;
+
+export type PreselectedShareType = Exclude<
+  z.infer<typeof preselectedResourceType>,
+  "slide deck" | "additional material"
+>;
+
+export type ShareResourceType = LessonShareSchema["type"];
 
 export const preselectedDownloadTypeMap: Record<
   PreselectedDownloadType,
@@ -35,6 +77,17 @@ export const preselectedDownloadTypeMap: Record<
   "exit quiz": ["exit-quiz-questions", "exit-quiz-answers"],
   worksheet: ["worksheet-pdf", "worksheet-pptx"],
   "additional material": ["supplementary-pdf", "supplementary-docx"],
+  all: "all",
+};
+
+export const preselectedShareTypeMap: Record<
+  PreselectedShareType,
+  LessonShareSchema["type"][] | "all"
+> = {
+  "starter quiz": ["intro-quiz-questions"],
+  "exit quiz": ["exit-quiz-questions"],
+  worksheet: ["worksheet-pdf"],
+  video: ["video"],
   all: "all",
 };
 
@@ -83,7 +136,7 @@ export const schema = z.object({
       message: "Accept terms and conditions to continue",
     }),
   }),
-  downloads: z
+  resources: z
     .array(z.string(), {
       errorMap: () => ({
         message: "Select at least one lesson resource to continue",
@@ -92,10 +145,10 @@ export const schema = z.object({
     .min(1),
 });
 
-export type DownloadFormValues = z.infer<typeof schema>;
+export type ResourceFormValues = z.infer<typeof schema>;
 
-export type DownloadFormProps = DownloadFormValues & {
-  onSubmit: (values: DownloadFormValues) => Promise<string | void>;
+export type ResourceFormProps = ResourceFormValues & {
+  onSubmit: (values: ResourceFormValues) => Promise<string | void>;
 };
 
-export type ErrorKeysType = keyof Omit<DownloadFormProps, "onSubmit">;
+export type ErrorKeysType = keyof Omit<ResourceFormProps, "onSubmit">;