+++ src/components/UnitModal/UnitModal.tsx
@@ -119,6 +119,14 @@ const UnitModal: FC<UnitModalProps> = ({
                 <CurriculumUnitDetails
                   threads={unitData.threads}
                   lessons={unitData.lessons}
+                  priorUnitDescription={
+                    unitData.connection_prior_unit_description
+                  }
+                  futureUnitDescription={
+                    unitData.connection_future_unit_description
+                  }
+                  priorUnitTitle={unitData.connection_prior_unit_title}
+                  futureUnitTitle={unitData.connection_future_unit_title}
                 />
               </Box>
             )}
@@ -201,6 +209,14 @@ const UnitModal: FC<UnitModalProps> = ({
                                     unitTitle: optionalUnit.title,
                                     threads: unitData.threads,
                                     lessons: optionalUnit.lessons,
+                                    priorUnitDescription:
+                                      optionalUnit.connection_prior_unit_description,
+                                    futureUnitDescription:
+                                      optionalUnit.connection_future_unit_description,
+                                    priorUnitTitle:
+                                      optionalUnit.connection_prior_unit_title,
+                                    futureUnitTitle:
+                                      optionalUnit.connection_future_unit_title,
                                   });
                                 }}
                               />